Swapped COO sparse arrays with BSR sparse arrays and changed sparse output types to reflect the underlying sparse arrays.
Previously the sparse output types were denoted as `NDArray` even when they were other types (e.g. `coo_array`).
Now the types are denoted with the true types returned (and accepted) by functions. Additionally, `coo_array` was
replaced with `bsr_array` for the purpose of allowing easy arithmetic operations with minimal user effort.
Finally, the `bsr_allclose` method was made as a helper function so that `np.allclose` is not used outside of its
scope, for example, when working with sparse matrices.